backend/pyproject.toml
[tool.poetry]
name = "agent-canvas-backend"
version = "0.1.0"
description = "Backend for Agent Canvas export and chat"
authors = ["Dev <dev@example.com>"]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"  # or ">=3.13,<3.14" if you want only 3.13
fastapi = ">=0.110,<1.0.0"
uvicorn = { extras = ["standard"], version = ">=0.31.1,<1.0.0" }
pydantic = ">=2.5,<3"
python-multipart = ">=0.0.9"
azure-ai-agents = "1.2.0b5"
agent-framework = "1.0.0b251007"
agent-framework-azure-ai = "1.0.0b251007"
azure-ai-projects = "^1.0.0b6"
azure-identity = "^1.19.0"
openai = "^1.57.0"
azure-core = "^1.30.0"
python-dotenv = "^1.0.0"
